minor styling, refactor our arrangeDepts func to new file
