+++ components/departments-sidebar/index.tsx
@@ -1,38 +1,8 @@
 import style from './style.module.css'
 import { map, isEmpty, filter } from 'lodash'
+import { arrangeDepartments } from './../utils/arrange-departments'
 
 const DepartmentsSidebar = ({ departments, onClick, departmentFilter }) => {
-  const findSubdepartments = (departmentsList, parentDepartmentId) => {
-    let subdepartments = []
-    map(departmentsList, (department) => {
-      if (department?.parent?.id === parentDepartmentId) {
-        const matchingSubdepartments = findSubdepartments(
-          departmentsList,
-          department.id
-        )
-
-        if (!isEmpty(matchingSubdepartments)) {
-          department.children = matchingSubdepartments
-        }
-        subdepartments.push(department)
-      }
-    })
-    return subdepartments
-  }
-
-  const arrangeDepartments = (departmentsList) => {
-    const parentDepartments = filter(departmentsList, (d) => !d.parent)
-    const organizedDepartments = map(parentDepartments, (parentDepartment) => {
-      const subDepartments = findSubdepartments(
-        departmentsList,
-        parentDepartment.id
-      )
-      return { ...parentDepartment, children: subDepartments }
-    })
-
-    return organizedDepartments
-  }
-  console.log(departmentFilter)
   const renderDepartmentListItems = (departmentsGroup) => {
     return (
       <ul>
