+++ components/people-cards/style.module.css
@@ -46,5 +46,5 @@
 }
 
 .personCardBody {
-  padding: 2px 16px;
+  padding: 1em;
 }
